+++ biopipen/scripts/tcr/Immunarch-overlap.R
@@ -89,8 +89,25 @@ do_one_case = function(name, case, ov_dir) {
 
     if (!is.null(case$analyses$cases) && length(case$analyses$cases) > 0) {
         for (aname in names(case$analyses$cases)) {
+            if (case$analyses$cases[[aname]]$method == "none") next
+
             print(paste0("    Analysis: ", aname))
-            imm_ova = repOverlapAnalysis(imm_ov, .method = case$analyses$cases[[aname]]$method)
+            k = min(n_samples - 1, 2)
+            tryCatch({
+                imm_ova = repOverlapAnalysis(
+                    imm_ov, .k = k, .perp = 1e-5, .method = case$analyses$cases[[aname]]$method
+                )
+            }, error = function(e) {
+                if (grepl("cmdscale", e)) {
+                    stop(paste0(
+                        "Too few samples (", n_samples, ") for overlap analyses.\n",
+                        "You can use a different method or set it to `none` to skip it.\n",
+                        "(Orginal error: ", e, ")"
+                    ))
+                } else {
+                    stop(e)
+                }
+            })
             avis_args = case$analyses$cases[[aname]]$vis_args
             avis_args$.data = imm_ova
             ap = do_call(vis, avis_args)